+++ euporie/__init__.py
@@ -8,3 +8,7 @@ __author__ = "Josiah Outram Halstead"
 __email__ = "josiah@halstead.email"
 __copyright__ = f"© 2022, {__author__}"
 __license__ = "MIT"
+
+from euporie.log import default_logs
+
+default_logs()